reorganize workers
