+++ imjoy/VERSION
@@ -1,4 +1,4 @@
 {
-    "version": "0.8.8",
+    "version": "0.8.9",
     "api_version": "0.1.2"
 }
